rename: component name Canvas to TaskBoard
